+++ src/nmath/bessel_i.c
@@ -64,7 +64,7 @@ double bessel_i(double x, double alpha, double expo)
                      : bessel_k(x, -alpha, expo) * ((ize == 1) ? 2. : 2. * exp(-2. * x)) / M_PI * sin(-M_PI * alpha)));
     }
     nb = 1 + (long)na; /* nb-1 <= alpha < nb */
-    alpha -= (nb - 1);
+    alpha -= (double)(nb - 1);
 #ifdef MATHLIB_STANDALONE
     bi = (double *)calloc(nb, sizeof(double));
     if (!bi)
@@ -80,7 +80,7 @@ double bessel_i(double x, double alpha, double expo)
             MATHLIB_WARNING4(_("bessel_i(%g): ncalc (=%ld) != nb (=%ld); alpha=%g. Arg. out of range?\n"), x, ncalc, nb,
                              alpha);
         else
-            MATHLIB_WARNING2(_("bessel_i(%g,nu=%g): precision lost in result\n"), x, alpha + nb - 1);
+            MATHLIB_WARNING2(_("bessel_i(%g,nu=%g): precision lost in result\n"), x, alpha + (double)nb - 1);
     }
     x = bi[nb - 1];
 #ifdef MATHLIB_STANDALONE
@@ -120,7 +120,7 @@ double bessel_i_ex(double x, double alpha, double expo, double *bi)
                                      sin(-M_PI * alpha)));
     }
     nb = 1 + (long)na; /* nb-1 <= alpha < nb */
-    alpha -= (nb - 1);
+    alpha -= (double)(nb - 1);
     I_bessel(&x, &alpha, &nb, &ize, bi, &ncalc);
     if (ncalc != nb)
     { /* error input */
@@ -128,7 +128,7 @@ double bessel_i_ex(double x, double alpha, double expo, double *bi)
             MATHLIB_WARNING4(_("bessel_i(%g): ncalc (=%ld) != nb (=%ld); alpha=%g. Arg. out of range?\n"), x, ncalc, nb,
                              alpha);
         else
-            MATHLIB_WARNING2(_("bessel_i(%g,nu=%g): precision lost in result\n"), x, alpha + nb - 1);
+            MATHLIB_WARNING2(_("bessel_i(%g,nu=%g): precision lost in result\n"), x, alpha + (double)nb - 1);
     }
     x = bi[nb - 1];
     return x;
